rename: 'promise' --> 'fsw-promise', add .js extension
