+++ utils/index.js
@@ -4,7 +4,7 @@ module.exports = {
     'createChangeDetail': require('./create-change-detail.js'),
     'resolvePathArray': require('./resolve-path-array.js'),
     'jsonSeperator': require('./json-separator.js'),
-    'Promise': require('./promise'),
+    'FSWPromise': require('./fsw-promise.js'),
     'FSWArray': require('./fsw-array.js'),
     '__proto__': null
 }
